lab2/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
bitarray==2.7.0
certifi==2022.12.7
charset-normalizer==3.0.1
cytoolz==0.12.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
importlib-resources==5.10.2
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
parsimonious==0.8.1
pkgutil-resolve-name==1.3.10
protobuf==3.19.5
py-solc-x==1.1.1
pycryptodome==3.17
pyrsistent==0.19.3
requests==2.28.2
rlp==2.0.1
semantic-version==2.10.0
six==1.16.0
toolz==0.12.0
urllib3==1.26.14
varint==1.0.2
web3==5.31.3
websockets==9.1
yarl==1.8.2
zipp==3.12.1
